src/third_party/WebKit/LayoutTests/fast/html/imports/rel-style-to-import-expected.txt

CONSOLE WARNING: line 18: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
CONSOLE WARNING: CSS cannot be loaded from `file:` URLs unless they end in a `.css` file extension.
Updating attribute value shouldn't result import loading.
Before update:
PASS target.sheet is non-null.
PASS target.import is null
After update:
PASS target.sheet is non-null.
PASS target.import is null
PASS successfullyParsed is true

TEST COMPLETE
